+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## 0.5.2 - 22-05-23
 
 ### Changed
 * Update FCS to 43.7.400-preview.23271.1
@@ -8,6 +8,7 @@
 ### Fixed
 * Wrong signature for member getter, setter with extra parameters. [#52](https://github.com/nojaf/telplin/issues/52)
 * Setter with different input than return type. [#61](https://github.com/nojaf/telplin/issues/61)
+* Types in a recursive module require the and keyword. [#62](https://github.com/nojaf/telplin/issues/62)
 
 ## 0.5.1 - 11-05-23